pyproject.toml
[build-system]
requires = ["setuptools>=67"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["rpt_dosi"]
include-package-data = true
zip-safe = false

[tool.setuptools.dynamic]
version = { file = "VERSION" }
readme = { file = ["README.md"] }

[project]
name = "rpt-dosi"
dynamic = ["version", "readme"]
authors = [{ name = "David Sarrut", email = "david.sarrut@creatis.insa-lyon.fr" }]
description = "RadioPharmaceutical Therapy Dosimetry"
license = { file = "LICENSE.md" }
keywords = ["html"]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Operating System :: OS Independent",
]
dependencies = [
    "colored",
    "tqdm",
    "pydicom",
    "click",
    "python-box<7.0.0",
    "numpy",
    "SimpleITK",
    "matplotlib",
    "python-Levenshtein",
    "bs4",
    "selenium",
    "radioactivedecay",
    "questionary",
    "pandas",
    "openpyxl",
    "gatetools"
]
requires-python = ">=3.8"

[project.scripts]

rpt_tests = "rpt_dosi.bin.rpt_tests:go"

rpt_crop_bg = "rpt_dosi.bin.rpt_crop_bg:go"
rpt_roi_crop = "rpt_dosi.bin.rpt_roi_crop:go"
rpt_spect_roi_statistics = "rpt_dosi.bin.rpt_spect_roi_statistics:go"
rpt_resample_ct = "rpt_dosi.bin.rpt_resample_ct:go"
rpt_resample_spect = "rpt_dosi.bin.rpt_resample_spect:go"
rpt_resample_roi = "rpt_dosi.bin.rpt_resample_roi:go"
rpt_image_info = "rpt_dosi.bin.rpt_image_info:go"
rpt_image_set_metadata = "rpt_dosi.bin.rpt_image_set_metadata:go"
rpt_spect_update = "rpt_dosi.bin.rpt_spect_update:go"

rpt_dose = "rpt_dosi.bin.rpt_dose:go"
rpt_dose_rate = "rpt_dosi.bin.rpt_dose_rate:go"
rpt_tmtv = "rpt_dosi.bin.rpt_tmtv:go"

opendose_web_get_isotopes_list = "rpt_dosi.bin.opendose_web_get_isotopes_list:go"
opendose_web_get_sources_list = "rpt_dosi.bin.opendose_web_get_sources_list:go"
opendose_web_get_svalues = "rpt_dosi.bin.opendose_web_get_svalues:go"

# new db scripts
rpt_db_info = "rpt_dosi.bin.rpt_db_info:go"

# old db scripts
rpt_db_rois_activity = "rpt_dosi.bin.rpt_db_rois_activity:go"
rpt_db_add_rois = "rpt_dosi.bin.rpt_db_add_rois:go"
rpt_db_tac_plot = "rpt_dosi.bin.rpt_db_tac_plot:go"
rpt_db_tac_triexpo = "rpt_dosi.bin.rpt_db_tac_triexpo:go"
rpt_db_set_spect_datetime_from_dicom = "rpt_dosi.bin.rpt_db_set_spect_datetime_from_dicom:go"
rpt_db_spect_calibration = "rpt_dosi.bin.rpt_db_spect_calibration:go"

rpt_dicom_browse = "rpt_dosi.bin.rpt_dicom_browse:go"
rpt_dicomdir_info = "rpt_dosi.bin.rpt_dicomdir_info:go"
rpt_dicom_select = "rpt_dosi.bin.rpt_dicom_select:go"
rpt_dicom_db = "rpt_dosi.bin.rpt_dicom_db:go"
